% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_parameters.R
\name{calc_df_eff}
\alias{calc_df_eff}
\title{Calculate the effective degrees of freedom}
\usage{
calc_df_eff(X, experimental_design)
}
\description{
The effective degrees of freedom are the sum of all non NA observation
minus the number of non-empty conditions.
}
\keyword{internal}
