% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posterior.R
\name{calculate_posterior}
\alias{calculate_posterior}
\title{Calculate the posterior distribution of the mean for each condition}
\usage{
calculate_posterior(X, mu0, sigma20, nu, eta, rho, zeta,
  experimental_design, niter = 1000, nchains = 4, ncores = nchains,
  batch_size = 1000, verbose = TRUE)
}
\description{
The method uses MCMC sampling to find a good approximation to the posterior
for each mean. The method is called using the parameters inferred from
\code{fit_hyperparameters()}. The parameter \code{niter} specifies how many
samples are drawn for each mean. The number of samples that are available
after burnin are \eqn{samples = niter / 2 * nchains}
}
