% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{dist_approx}
\alias{dist_approx}
\title{Distance matrix computation respecting nonrandom missing values}
\usage{
dist_approx(X, mu_mis = NULL, var_mis = NULL, experimental_design = NULL,
  hyperparameters = NULL, ...)
}
\arguments{
\item{X}{the transposed (!) data matrix}

\item{mu_mis}{mean of the replacement values. Can be a single number, a
vector with one number for each sample or a matrix with the same dimensions
as X.}

\item{var_mis}{variance of the replacement values. Can be a single number, a
vector with one number for each sample or a matrix with the same dimensions
as X.}

\item{experimental_design}{if no good replacement values are available
you can provide the experimental_design vector}

\item{hyperparameters}{a list with four elements (mu0, sigma20, rho and zeta),
that are used to calculate a good replacement value.}
}
\description{
The method calculates the distances between each row of the matrix X and
returns a distance matrix. Because of the missing values no exact distance can
be calculated instead realistic values for the missing values are considered
and the mean with the corresponding variance is calculated for each distance.
}
\details{
If no appropriate replacement values are available and the method is called
without any parameter in addition to X, it calls `fit_hyperparameters` and
infers them.
}
\seealso{
stats::dist
}
