% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{find_approx_for_missing}
\alias{find_approx_for_missing}
\title{Find a Gaussian approximation for each missing value in X}
\usage{
find_approx_for_missing(X, params = NULL, experimental_design, mup,
  sigma2mup, sigma2p, rho, zeta)
}
\value{
a list with with two elements `mu_mis` and `var_mis`
  both of which are matrices with the same dimensions as X
  and values wherever X has a missing value.
}
\description{
Find a Gaussian approximation for each missing value in X
}
