% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_parameters.R
\name{fit_feature_mean_uncertainties}
\alias{fit_feature_mean_uncertainties}
\title{Calculate the variance of the mean estimator}
\usage{
fit_feature_mean_uncertainties(X, rho, zeta, nu, eta, mu0, sigma20,
  experimental_design)
}
\description{
Calculate the variance of the mean estimator
}
