% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_parameters.R
\name{fit_feature_means}
\alias{fit_feature_means}
\title{Estimate the mean for each feature}
\usage{
fit_feature_means(X, sigma2p, sigma2mus, rho, zeta, nu, eta, mu0, sigma20,
  experimental_design)
}
\description{
Estimate the mean for each feature
}
