% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_parameters.R
\name{fit_feature_variances}
\alias{fit_feature_variances}
\title{Find the most likely variance explaining the values for each row in X}
\usage{
fit_feature_variances(X, mup, rho, zeta, nu, eta, experimental_design,
  upper = 1000)
}
\description{
Find the most likely variance explaining the values for each row in X
}
