% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dropout_curves.R
\name{fit_global_dropout_curves}
\alias{fit_global_dropout_curves}
\alias{fit_sample_dropout_curves}
\alias{fit_global_scale_dropout_curves}
\title{Find the inverse probit parameters that best explain the pattern of missing
values.}
\usage{
fit_global_dropout_curves()

fit_sample_dropout_curves()

fit_global_scale_dropout_curves(X, mup, sigma2p, experimental_design, prev_zeta,
  prev_rho, max_iter = 10, epsilon = 1e-07)
}
\description{
There are three different modes for fitting the inverse probit dropout
curves:
\describe{
\item{fit_global_dropout_curves}{One curve for all samples}
\item{fit_sample_dropout_curves}{One curve for each sample}
\item{fit_global_scale_dropout_curves}{One curve for each sample,
  but all have a common scale}
}
}
