% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hyperparameters.R
\name{fit_hyperparameters}
\alias{fit_hyperparameters}
\title{Iteratively update the hyperparameter until convergence}
\usage{
fit_hyperparameters(X, experimental_design,
  dropout_curve_calc = c("global", "global_scale", "sample"),
  frac_subsample = 1, n_subsample = round(nrow(X) * frac_subsample),
  max_iter = 10, epsilon = 0.001, verbose = FALSE)
}
\description{
Iteratively update the hyperparameter until convergence
}
