% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hyperparameters.R
\name{fit_location_prior}
\alias{fit_location_prior}
\title{Fit a Gaussian location prior over all samples}
\usage{
fit_location_prior(X, mup, zeta, rho, experimental_design)
}
\value{
list with elements mu0 and sigma20
}
\description{
The function takes the regularized feature means. The global mu0 is the
mean of the regularized feature means. It then calculates the unregularized
feature means right of the global mean (so that missing values are less
problematic). The global variance estimate is the variance of those
unregularized values to the global mean.
}
