% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_parameters.R
\name{fit_unregularized_feature_means}
\alias{fit_unregularized_feature_means}
\title{Calculate the unregularized means for all rows for which mup is larger than mu0}
\usage{
fit_unregularized_feature_means(X, mup, mu0, zeta, rho,
  experimental_design)
}
\description{
Calculate the unregularized means for all rows for which mup is larger than mu0
}
\keyword{internal}
