% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hyperparameters.R
\name{fit_variance_prior}
\alias{fit_variance_prior}
\title{Fit a inverse Chi-squared variance prior}
\usage{
fit_variance_prior(X, rho, zeta, experimental_design)
}
\description{
Run maximum likelihood estimate on the density of the F distribution with
the unregularized variance estimates.
}
\keyword{internal}
