% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate a data set according to the probabilistic dropout model}
\usage{
generate_synthetic_data(n_rows, n_replicates,
  n_conditions = length(n_replicates), frac_changed = 0.1,
  n_changed = round(n_rows * min(1, frac_changed)),
  experimental_design = c(unlist(mapply(rep, seq_len(n_conditions),
  n_replicates))), mu0 = 20, sigma20 = 10, nu = 3, eta = 0.3,
  rho = rep(18, times = if (length(n_replicates) == 1) n_replicates *
  n_conditions else sum(n_replicates)), zeta = rep(-1, times = if
  (length(n_replicates) == 1) n_replicates * n_conditions else
  sum(n_replicates)))
}
\arguments{
\item{n_rows}{integer. The number of rows in the new dataset}

\item{n_replicates}{integer or vector. The number of replicates in each
condition}

\item{n_conditions}{}
}
\description{
Generate a data set according to the probabilistic dropout model
}
