% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic_dropouts.R
\name{match_probdropout_skewed_normal}
\alias{match_probdropout_skewed_normal}
\title{Match a skewed normal distribution to the probabilistic dropout density}
\usage{
match_probdropout_skewed_normal(mu, sigma2, rho, zeta)
}
\arguments{
\item{mu}{double. The mean of the observed values.}

\item{sigma2}{double. The variance of the mu estimate.}

\item{rho}{vector. The positions of the inflection points of the dropout
sigmoids for each sample. Can also be a single number that is repeated
\code{nmis} times. Defaults to an empty vector.}

\item{zeta}{vector.  The flattness of the dropout sigmoids for each sample.
Can also be a single number that is repeated \code{nmis} times.
Defaults to an empty vector.}
}
\description{
Find the mode of probdropout density and take 5 points around the mode
with a distance according to the Laplace approximation variance and calculate
the parameters of the skewed normal that most closely matches those
5 points.
}
