% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic_dropouts.R
\name{mode_probdropout}
\alias{mode_probdropout}
\title{Find the mode of the probabilistic dropout distribution function}
\usage{
mode_probdropout(mu, sigma2, rho, zeta)
}
\description{
Find the mode of the probabilistic dropout distribution function
}
\section{Warning}{

This function is not vectorized
}

