% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_feature_parameters.R
\name{predict_feature_parameters}
\alias{predict_feature_parameters}
\title{Find feature parameters for elements of X with a fixed set of hyperparameters}
\usage{
predict_feature_parameters(X, experimental_design, hyper_params = NULL,
  mu0, sigma20, nu, eta, rho, zeta, mup = NULL, sigma2p = NULL,
  sigma2mup = NULL, max_iter = 10, epsilon = 0.001,
  verbose = FALSE)
}
\arguments{
\item{X}{data matrix}

\item{experimental_design}{a vector that specifies which samples belong
to the same condition.}

\item{hyper_params}{a list with 5 elements (`eta`, `nu`, `mu0`, `sigma20`,
`rho`, and `zeta`). Alternatively the `hyper_params` can be specified
individually.}

\item{mu0}{the global mean around which the row means are drawn}

\item{sigma20}{the global variance specifying the spread of means around `mu0`.}

\item{nu}{degrees of freedom for the the global variance prior.}

\item{eta}{scale of the global variance prior.}

\item{rho}{vector specifying the intensity where the chance of a dropout is
50/50. Length is either one or \code{ncol(X)}.}

\item{zeta}{vector specifying the scale of the dropout curve.
Length is either one or \code{ncol(X)}.}

\item{mup}{Optional matrix that fixes the mean for each row and condition. Default `NULL`}

\item{sigma2p}{Optional vector that fixes the variance for each row. Default `NULL`}

\item{sigma2mup}{Optional matrix that fixes the uncertainty of the mean
for each row and condition. Default `NULL`.}

\item{max_iter}{the maximum number of iterations. Default: 10}

\item{epsilon}{the error under which the result is considered converged.
Default: 0.001}

\item{verbose}{boolean that indicates if verbose output is printed to the console.}
}
\value{
list with three elements
  \describe{
    \item{mup}{a matrix with size \code{nrow(X) * unique(experimental_design)}
      with the means for each feature}
    \item{sigma20}{a numeric vector with the variance for each feature}
    \item{sigma2mup}{a matrix with size \code{nrow(X) * unique(experimental_design)}
      with the uncertainty for each `mup`}
  }
}
\description{
Find feature parameters for elements of X with a fixed set of hyperparameters
}
