% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodd_parameters.R
\name{format.prodd_parameters}
\alias{format.prodd_parameters}
\alias{print.prodd_parameters}
\title{Collection of parameters generated when fitting the probabilistic dropout model}
\usage{
\method{format}{prodd_parameters}(x, ...)

\method{print}{prodd_parameters}(x, ...)
}
\arguments{
\item{x}{an object of calss  prodd_parameters}

\item{...}{additional arguments to be passed to or from the method}
}
\description{
Collection of parameters generated when fitting the probabilistic dropout model
}
