% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_posterior.R
\name{sample_protein_means}
\alias{sample_protein_means}
\title{Estimate the mean of each protein and condtion using MCMC sampling.}
\usage{
sample_protein_means(X, params, niter = 1000, nchains = 4,
  ncores = nchains, batch_size = 1000, verbose = TRUE, ...)
}
\arguments{
\item{X}{the numerical data where each column is one sample and each row
is one protein. Missing values are coded as \code{NA}.}

\item{params}{an object of class `prodd_parameters`}

\item{niter}{the number of iteration for each posterior and chain. In
the end you will have \code{niter * nchains / 2} samples. Default: 1000}

\item{nchains}{the number of chains to run in parallel. Default: 4}

\item{ncores}{the number of cores that are used in parallel. Deafault: nchains.}

\item{batch_size}{Often it is faster to run the inference for multiple
proteins at the same time, but there is a limit after which too many proteins
slow down the inference. Set the number of proteins that are considered in
a single run. Default: 1000}

\item{verbose}{boolean indicating how much output the function generates. D
Default: `TRUE`}

\item{...}{additional parameters for passed to \code{rstan::sampling}.}
}
\value{
a list with one matrix per condtion. Each matrix has one row
  per protein and one column per MCMC sample.
}
\description{
The method uses MCMC sampling implemented by Stan to find a good approximation
to the posterior for each mean. The method is called using the parameters inferred
from \code{fit_parameters()}. The parameter \code{niter} specifies how many
samples are drawn for each mean. The number of samples that are available
after burnin are \eqn{samples = niter / 2 * nchains}.
}
