% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic_dropouts.R
\name{variance_probdropout}
\alias{variance_probdropout}
\title{Find the variance of  of the probabilistic dropout distribution function}
\usage{
variance_probdropout(mu, sigma2, rho, zeta, log = FALSE, approx = TRUE)
}
\arguments{
\item{approx}{boolean. If TRUE match the probabilistic dropout
distribution to a skewed normal distribution and use its mean as an
approximation. If FALSE calculate the first moment using numerical
integration}
}
\description{
Find the variance of  of the probabilistic dropout distribution function
}
\section{Warning}{

This function is not vectorized
}

