% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{dist_approx}
\alias{dist_approx}
\title{Distance matrix computation respecting non-random missing values}
\usage{
dist_approx(X, params = NULL, by_sample = TRUE, mu_mis = NULL,
  var_mis = NULL)
}
\arguments{
\item{X}{the numerical data where each column is one sample and each row
is one protein. Missing values are coded as \code{NA}.}

\item{params}{an object of class `prodd_parameters` which for example is returned
by the \code{fit_parameters()} function.}

\item{by_sample}{boolean. Indicate if the distances between samples (columns) or
proteins (rows) is calculated. Default: TRUE.}

\item{mu_mis}{mean of the replacement values. Can be a single number, a
vector with one number for each sample or a matrix with the same dimensions
as X. Can be provided instead of the `params` parameters.}

\item{var_mis}{variance of the replacement values. Can be a single number, a
vector with one number for each sample or a matrix with the same dimensions
as X. Can be provided instead of the `params` parameters.}
}
\value{
a list with two elements.
  \describe{
    \item{mean}{a distance matrix with the mean of the distance estimate}
    \item{var}{the corresponding uncertainty to each distance estimate}
  }
}
\description{
The method calculates the distances between each samples or proteins of the matrix X and
returns a distance matrix and the corresponding uncertainty. Because of the missing
values no exact distance can be calculated instead realistic values for the missing
values are considered and the mean with the corresponding variance is
calculated for each distance.
}
\details{
Usually the method is called with the data matrix `X `and
the object that is returned by \code{fit_parameters()} or the result
obtained by first calling \code{transform_parameters()} to remove the group
information. The `params` object must be  of type `prodd_parameters`.

If particular information are available, where the missing values would have
been (ie. mean and variance for each missing value), they can instead of the
params object be provided in form of two matrices (with the same dimensions as X)
or individual values.

Unlike the \code{stats::dist} function which always calculates the distance
between the rows of the matrix and one transposes X to find the distances
between columns, this method uses the \code{by_sample} parameter. In `X` (and
correspondingly for `mu_mis` and `var_mis`)
the columns always correspond to the samples and the rows to the proteins. By default
the distances are calculated between the samples, to calculate the distances
between the proteins set \code{by_sample=FALSE}.
}
\seealso{
\code{\link{stats::dist}}
}
