% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic_dropouts.R
\name{mode_probdropout}
\alias{mode_probdropout}
\title{Find the mode of the probabilistic dropout distribution function}
\usage{
mode_probdropout(mu, sigma2, rho, zeta)
}
\arguments{
\item{mu}{double. The mean of the observed values.}

\item{sigma2}{double. The variance of the mu estimate.}

\item{rho}{vector. The positions of the inflection points of the dropout
sigmoids for each sample. Can also be a single number that is repeated
\code{nmis} times. Defaults to an empty vector.}

\item{zeta}{vector.  The flattness of the dropout sigmoids for each sample.
Can also be a single number that is repeated \code{nmis} times.
Defaults to an empty vector.}
}
\description{
Find the mode of the probabilistic dropout distribution function
}
\section{Warning}{

This function is not vectorized
}

