% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_test.R
\name{resampling_test}
\alias{resampling_test}
\title{Test the difference between x and y using resampling.}
\usage{
resampling_test(x, y, nmax = length(x) * length(y),
  alternative = c("two.sided", "less", "greater"), mu = 0)
}
\arguments{
\item{x}{first numeric vector}

\item{y}{second numeric vector}

\item{nmax}{the maximum number of comparisons. Can be used if the full
precision is not needed. Defaults to making all possible comparisons}

\item{alternative}{how are x and y compared. Same principles as
\code{t.test}}

\item{mu}{test if there is at least distance mu between x and y.
Equivalent to calling \code{resampling_test(x, y + mu)}}
}
\description{
The method is an alternative to \code{mean(outer(x, y, `<`))}
which doesn't allocate memory and is thus much faster.
}
