% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilistic_dropouts.R
\name{variance_probdropout}
\alias{variance_probdropout}
\title{Find the variance of  of the probabilistic dropout distribution function}
\usage{
variance_probdropout(mu, sigma2, rho, zeta, log = FALSE, approx = TRUE)
}
\arguments{
\item{mu}{double. The mean of the observed values.}

\item{sigma2}{double. The variance of the mu estimate.}

\item{rho}{vector. The positions of the inflection points of the dropout
curves for each sample. Can also be a single number that is repeated
\code{nmis} times. Defaults to an empty vector.}

\item{zeta}{vector.  The scale of the dropout curves for each sample.
Can also be a single number that is repeated \code{nmis} times.
Defaults to an empty vector.}

\item{log}{boolean if the log of the density is returned.}

\item{approx}{boolean. If TRUE match the probabilistic dropout
distribution to a skewed normal distribution and use its variance as an
approximation. If FALSE calculate the second moment using numerical
integration}
}
\description{
Find the variance of  of the probabilistic dropout distribution function
}
\section{Warning}{

This function is not vectorized
}

